force int

+++ app/TypiCMS/Modules/News/Controllers/Admin/NewsController.php
@@ -103,7 +103,7 @@ class NewsController extends BaseController {
 
 		// add checkboxes data
 		foreach (Config::get('app.locales') as $locale) {
-			$data[$locale]['status'] = Input::get($locale.'.status');
+			$data[$locale]['status'] = (int) Input::get($locale.'.status');
 		}
 
 		if ( $this->form->update( $data ) ) {
